# torque-api/api/openapi.yaml
openapi: 3.1.0
info:
  title: Torque API
  version: 0.52.0
  description: |
    Torque CRM backend. Conventions (ADR-001): wire uses snake_case; frontend
    transforms to camelCase. Session is carried in httpOnly cookies; never
    ship a Bearer token.

    **S52 refresh.** Covers endpoint families shipped through S51 —
    leads, pipes, confirmations, meetings, tasks, inbox, templates,
    campaigns, workflows, agents (+ playground/metrics/knowledge/TTS),
    billing, quotas, integrations (Google/TinyERP/Meta/SZChat), public
    webhooks (billing + lead), analytics, performance, master. Shapes
    are intentionally compact — the wire details live in the respective
    handler files; this spec is the map, not the manual.
  contact:
    name: Torque engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.torque.app
    description: Production
...
paths:
...
  /api/v1/workflows:
    get:
      tags: [workflows]
      operationId: listWorkflows
      summary: List workflows
      security: [{ cookieAuth: [] }]
      responses:
        "200": { description: OK }
    post:
      tags: [workflows]
      operationId: createWorkflow
      summary: Create workflow draft
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      responses:
        "201": { description: Created }
  /api/v1/workflows/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [workflows]
      operationId: getWorkflow
      summary: Workflow detail (with steps)
      security: [{ cookieAuth: [] }]
      responses:
        "200": { description: OK }
    patch:
      tags: [workflows]
      operationId: updateWorkflow
      summary: Update workflow metadata
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      responses:
        "200": { description: Updated }
  /api/v1/workflows/{id}/steps:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [workflows]
      operationId: createWorkflowStep
      summary: Add a step to a workflow
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      responses:
        "201": { description: Created }
  /api/v1/workflows/{id}/steps/{stepId}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: stepId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [workflows]
      operationId: updateWorkflowStep
      summary: Replace step config
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      responses:
        "200": { description: Updated }
    delete:
      tags: [workflows]
      operationId: deleteWorkflowStep
      summary: Remove step
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      responses:
        "204": { description: Deleted }
  /api/v1/workflows/{id}/runs:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [workflows]
      operationId: enqueueWorkflowRun
      summary: Enqueue a debug run
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      responses:
        "202": { description: Enqueued }
  /api/v1/runs/{id}/steps:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [workflows]
      operationId: listRunSteps
      summary: Trace timeline for a run (refetch 2s while pending|running)
      security: [{ cookieAuth: [] }]
      responses:
        "200": { description: OK }

  # ---------- Agents / Copilot (F06) --------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __torque_session
  parameters:
    CSRFHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
      description: Must echo the __torque_csrf cookie (double-submit).
...